Add cloud module to project
Added support for 'cloud' to the user project, to enable cloud related functionalities. The makefile (`CMakeLists.txt`) and the source code (`applications/main.c`) have been updated with the new `cloud` module. Along with the addition of the 'cloud' folder which contains 'tencent_cloud.c'  that connects the project with the Tencent cloud. In `tencent_cloud.c`, The Tencent cloud related functionalities, mostly MQTT services, are added for operations on the cloud. Also, new properties and events for the Tencent cloud were set-up in 'data_config.c'.
The update also included the organization of other files in the project for cleaner code.